vim(folding): disable nvim-ufo and use built-in treesitter foldtext
nvim-ufo is broken on recent configuration with latest Nvim core or
other plugin versions (e.g. nvim-treesitter), possibly because of some
backward-incompatible API changes regarding MetaNode. There does not
seem to be a proper fix in nvim-ufo at this point. So we completely
opt out of the nvim-ufo plugin and switch to built-in foldtext available
since Nvim 0.10.0.

The only benefit nvim-ufo was providing besides highlighted foldtext is
a more fine-grained control in foldtext (e.g. how many lines are being
folded), but this is not critical to me. Broken configs and errors are
more annoying.

Change-Id: I4bc88fd9bf1a1915b0a00aa92c1df74198ffebad