vim: pin mason plugin versions at 1.x
mason.nvim and mason-lspconfig.nvim now turned into v2.0 with a few
breaking changes that are incompatible with the current config (and
older neovim versions).

For now we'll have to pin these plugin versions at a legacy one (1.x).
It seems that under some environments lazy.nvim fail to check out the
correct version as specified (but latest HEAD) for some reason, but
we'll try to figure it out soon.