bin/pbcopy: do not use osc52 on a local machine (e.g. mac)
This fixes a bug where copying to the system clipboard on neovim on a
local machine (outside TMUX) did not work as expected.

This custom pbcopy script, which was intended for remote terminal
environments, would be picked (instead of system pbcopy) by neovim's
clipboard provider, but if the standard output consisting of a OSC52
control sequence is *captured* rather than transmitted to the terminal,
nothing would have been copied.